# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Expo AHRO Telecoms",
        "description": "Documentación de la API para Expo AHRO",
        "contact": {
            "email": "tu-email@dominio.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/visitantes": {
            "get": {
                "tags": [
                    "Visitantes"
                ],
                "summary": "Obtener todos los visitantes",
                "operationId": "a1434dcd119a6bda3dd1b9c34e3c76c3",
                "responses": {
                    "200": {
                        "description": "Lista de visitantes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Visitante"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Visitantes"
                ],
                "summary": "Crear un nuevo visitante",
                "operationId": "942bc27c2cdc1dc9ec16a7dda004ee25",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "nombre_completo": {
                                        "description": "Nombre completo del visitante",
                                        "type": "string"
                                    },
                                    "correo": {
                                        "description": "Correo electrónico del visitante",
                                        "type": "string"
                                    },
                                    "telefono": {
                                        "description": "Teléfono del visitante",
                                        "type": "string"
                                    },
                                    "productos": {
                                        "description": "Productos asociados al visitante",
                                        "type": "array",
                                        "items": {
                                            "description": "ID del producto",
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Visitante creado"
                    },
                    "422": {
                        "description": "Validación de datos fallida"
                    }
                }
            }
        },
        "/api/visitantes/{id}": {
            "get": {
                "tags": [
                    "Visitantes"
                ],
                "summary": "Obtener un visitante por ID",
                "operationId": "dd21b35ef6c2e60cf8c0a0972431275d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del visitante",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Información del visitante",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Visitante"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Visitante no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Visitantes"
                ],
                "summary": "Actualizar un visitante",
                "operationId": "e99a6651d4f53d6493fb12394d0ec8bf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del visitante",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "nombre_completo": {
                                        "type": "string"
                                    },
                                    "correo": {
                                        "type": "string"
                                    },
                                    "telefono": {
                                        "type": "string"
                                    },
                                    "productos": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Visitante actualizado"
                    },
                    "404": {
                        "description": "Visitante no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Visitantes"
                ],
                "summary": "Eliminar un visitante",
                "operationId": "2ad586253fece0ddae9b5c674a007ee3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del visitante",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Visitante eliminado"
                    },
                    "404": {
                        "description": "Visitante no encontrado"
                    }
                }
            }
        },
        "/api/visitantes/asignar-productos": {
            "post": {
                "tags": [
                    "Visitantes"
                ],
                "summary": "Asignar productos de interés a un visitante",
                "operationId": "c59e176f815a1d7cd741cf02700d499f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "visitante_id",
                                    "productos"
                                ],
                                "properties": {
                                    "visitante_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "productos": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Productos asignados correctamente"
                    },
                    "422": {
                        "description": "Validación de datos fallida"
                    }
                }
            }
        },
        "/api/visitantes/{visitante_id}/productos": {
            "get": {
                "tags": [
                    "Visitantes"
                ],
                "summary": "Obtener productos de interés de un visitante",
                "operationId": "e61633fcc17da5fb3ecc1dbae33f9a6e",
                "parameters": [
                    {
                        "name": "visitante_id",
                        "in": "path",
                        "description": "ID del visitante",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de productos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Producto"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Visitante no encontrado"
                    }
                }
            }
        },
        "/api/visitantes/producto/{producto_id}": {
            "get": {
                "tags": [
                    "Visitantes"
                ],
                "summary": "Obtener visitantes interesados en un producto",
                "operationId": "215dc9cf9ef3cf7c0093086322a61fda",
                "parameters": [
                    {
                        "name": "producto_id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de visitantes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Visitante"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Producto": {
                "title": "Producto",
                "required": [
                    "id",
                    "nombre"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Producto A"
                    },
                    "descripcion": {
                        "type": "string",
                        "example": "Descripción del producto A"
                    }
                },
                "type": "object"
            },
...
            "Visitante": {
                "title": "Visitante",
                "required": [
                    "id",
                    "nombre_completo",
                    "correo",
                    "telefono"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre_completo": {
                        "type": "string",
                        "example": "Juan Pérez"
                    },
                    "correo": {
                        "type": "string",
                        "format": "email",
                        "example": "juan.perez@example.com"
                    },
                    "telefono": {
                        "type": "string",
                        "example": "+521234567890"
                    },
                    "empresa": {
                        "type": "string",
                        "example": "Empresa S.A."
                    },
                    "cargo": {
                        "type": "string",
                        "example": "Gerente de ventas"
                    },
                    "fecha_registro": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-02T15:30:00Z"
                    },
                    "estado": {
                        "type": "string",
                        "example": "Nuevo"
                    },
                    "notas": {
                        "type": "string",
                        "example": "Cliente interesado en producto X"
                    }
                },
                "type": "object"
            }
        }
    },
...
}